+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2021/05/05 - v2.4.0: enhancement release
+
+1. Add domain to username impersonation in Agave.
+
+
 ## 2021/03/24 - v2.3.0: maintenance release
 
 1. Remove old test suite.